Avoid hoisting of local classes out of method bodies.
Doing so is unsound. We instead approximate local classes by their parents, but
only if the expected type is not fully defined.

This makes the test t2421_delitedsl1.scala in the commit pass. The oter test,
blockEscapesNeg.scala is modified to fail. Previously it failed outright but with the
new rules the nested class Bar is approximated to Object. That means that the block
containing `Foo.Bar` typechecks, but with type `Object` instead of the unreachable `Bar`.