Always ignore type in selectionProto
The type of a SelectionProto needs to be ignorable because there might be an
implicit conversion on the selection. E.g.

    implicit def a2b(x: A): B = ???
    val x: { a: A } = ???
    val b: B = x.a

This was previously handled by allowing implicit conversions in compatibility checks.
But it turns out we can afford to ignore the type of a selectProto and unignore on ambiguities
later.