Semantic printing
When refined printing a DefTree, if one is after typer,
print the symbol's definition plus the rhs of the tree
rather than the contents of the tree.

Why: After typer, it's always the symbol's contents that should matter.

Question: Can we enforce through the types that parts of DefTree reflected
by the symbol disappear/are ignored?