[backport] Fixes first sentence delimiters
Scaladoc places the first sentence in the method summary table and index. The first sentence is defined as the sequence of chars up to the first period (not as in JavaDoc where the sentence ends with the first period *followed by a blank*). As a consequence, the clause starting with `i.e.` is interpreted as sentende delimiter. This request replaces `i.e.` with `i&period;e&period;`. Alghough a valid HTML code, I do not know whether this change is compatible with other tools. And I assume that this is not the only source file affected.

(cherry picked from commit 43a8b74b6839d24dc2590fcf381c442e6d96ac30)