Fixes cut sentences in the generated scaladocs
This commit changes all first sentences of library functions which
contain `i.e.` or `e.g.` and adds a `,` to prevent that the scaladoc
summary sentence is cut after this abbreviation.

This is possible as pull/3824 fixed how Scaladoc parses the first
sentence of a method description into a sumary sentence(now the first
sentence has to end with a dot followed by whitespace).

Only docs in the core library are changed (src/library/**/*.scala)
and only if they occur in the first sentence.

Review by @heathermiller