SI-8787 Update doc for Regex
Simplify the exposition and update the examples.

Emphasize pattern matching, especially using `r.unanchored`
instead of `for (r(x,y,z) <- r findFirstIn text)`.

Certain details are moved to method docs.

It would be nice to fix matching package doc,
but the doc must attach to the package object, it seems.
Introducing a package object is not binary-compatible.

Includes a doc line edit on 2.12, anticipating the merge.