Make ElimByName change owners of by-name parameters
By-name parameters did not have their owner changed before.
This was not noticed in treecheck because the method generated
for the closure did not have its Method flag set by accident, so
owmer checking ignored the new val. Once the Methgod flag was set,
owner checking failed. Once changeOwner was added, a whole lot
of other things failed, which led to the fixes in the previous commits.