Rebinding ThisTypes
ThisTypes do escape. It seems that Scala 2 pickling produces ThisTypes that refer to base
classes of the class being compiled. If the base class is in a separate compilation unit,
this can lead to a stale class symbol in the ThisType. We solve this here by rebinding the
class symbol in the ThisType. We should also explore the alternative:

     class ThisType(tref: TypeRef) ...

That would do the rebinding as part of the generation denotation resolution mechanism.