Avoid cyclic reference errors in ShowScala
This is done by predicating some shortcuts on tp.symbol.isStatic
where tp is a NamedType with the condition tp.denotationIsCurrent,
i.e. we avoid forcing the denotation. This safe because the branches
taken on isStatic are optimizations.

This commit contains the minimum set of changes to make showScala pass.