Generalize equivalentThisTypes
       // We treat two prefixes A.this, B.this as equivalent if
       // A's selftype derives from B and B's selftype derives from A.

It makes sense to apply this principle to comparing this types in general, and not
only when being the prefix of a named type. Not doing so revealed a problem in
retyping Applications.scala after outerAccessors. There two type applications of
the same class had as prefixes (1) Applications.this, (2) Typer.this & Applications.this.
Typer and Applications match the principle, so the prefixes should be regarded as equivalent.
Why this manifested itself only after outerAccessors is an unsolved puzzle.