tpd.TreeOps#select creates non member syms after erasure.
Without this patch, ast.TreeInfo does not compile wfter next commit "Erasure should keep prefixes in
TermRefs". The reason is that outer path genertion generates a data race between

 - an outer accessor created in ExplicitOuter
 - a copy of the accessor created in ElimByName in a TreeTypeMap copy.

The patch avoids the data race by using a fixed symbol for the new TermRef generated in
outer.path during erasure.

But the fix seems ad-hoc. We should try to attack the problem at the root in TypeTreeMap.
The root problem seems to be that TypeTreeMap generates a copy of symbols whnMemere the copies
of the symbols give the same named type as the originals. We can either label the copied symbols
so that any types created from them become instances of WithNonMemberSym (which probably should be
renamed then). Or else we have a "linear" version of TreeMap that overrides sym denotations instead
of creating new symbols.