More refined printing options
When using RefinedPrinter we now have a choice whether when
printing a definition tree such as

    def foo(x: T): U

we print the parameter and result type info found in the tree or in the symbol.
Previously, we printed the sym info when after typer and the tree info before.
This turns out to be too inflexble. With the patch, we print the sym info if
option -Yprint-syms is set, and the tree info otherwise.

Also, align -Yno-deep-subtypes from camelCase to standard hyphenated option notation.

Tweak where unique ids are printed.