Add dotty repl & type stealer
Dotty requires a mangled bootclasspath to start.
It means that `console` mode of sbt doesn't work for us.
At least I wasn't able to make sbt fork in console,
so instead I've added a Scala-repl into dotty itself :-)

It would be good to make it use dotty one day when we have
a backend :-)