Avoid hoisting local classes
The patch disables hoisting of classes local to a block into the
result type of the block.

Instead, we widen the result type of the block to one which reflects
all refinements made to the parents type of the local class.

Test cases in avoid.scala, t1569.scala.

The original t1569.scala no longer works. Why is explained in neg/t1569-failedAvoid.scala