Better tests and bugfix for named args
The original problem was that in an expression

     f(x = bar(y = z))

only the outer named arg was eliminated by FirstTransform.

The first error was that the postcondition in FirstTransform did not get to the named arg, because
it got called from the overrdden typed method in TreeChecker, yet function arguments were evaluated
with typedUnadapted.

action: change Retyper and TreeChecker to override typedUndapped instead of typed.

This flushed out the second error: transformOther in FirstTransform needs to recursively transform the argument
of a NamedArg, because the framework itself does not handle NamedArg nodes.

Now, all tests pass except that TreeChecker itself fails -Ycheck:gettersSetters due to a problem
with handling by-name function types. This should be fixed in a separate PR.