don't try to make a body for a setter of a Java field
transformSym explicitly checks that a field is JavaDefined and does not create a symbol for it.

Creation of a setter body looks for the symbol and fails because it does not find it.

We do not need setter bodies for Java fields because we are not generating bytecode for them.