Fixes in TypeComparer for RefinedTypes.
The previous scheme did not propagate bounds correctly. More generally,
given a comparison

     T { X <: A }  <:   U { X <: B }

it would errenously decompose this to

     T <: U, A <: B

But we really need to check whether the total constraint for X in T { X <: A }
subsumes the total constraint for X in T { X <: B }

The new scheme propagates only if the binding in the lower type is an alias.
E.g.

    T { X = A }  <:   Y { X <: B }

decomposes to

    T { A = A }  <:   U, A <: B

The change uncovered another bug, where in the slow path we too a member relative to a refined type;
We need to "narrow" the type to a RefinedThis instead. (See use of "narrow" in TypeComparer).
That change uncovered a third bug concerning the underlying type of a RefinedThis. The last bug was fixed in a previous commit (84f32cd814f2e07725b6ad1f6bff23d4ee38c397).

Two tests (1048, 1843) which were pos tests for scalac but failed compling in dotc have
changed their status and location. They typecheck now, but fail later. They have been
moved to pending.

There's a lot of diagnostic code in TypeComparer to figure out the various problems. I left it in
to be able to come back to the commit in case there are more problems. The checks and diagnostics
will be removed in a subsequent commit.