SI-8691  SeqView throws exception when prepending a collection
Prepend was throwing segfaults as it was handled differently from Append.  This brings the two into line with each other.

There are various optimizations that could be applied that have not been, however, such as intercepting sequential prepends and generating one multi-prepend instead of nested single-element prepends.

Unit test added to verify minimally that bug is gone.