Fix returns from polymorphic methods
Return proto type was wrong; it was the type of the method
(with PolyParams if that type is polymorphic), where it should
have been the return type instantiated with the local type parameters.
We now instantiate the prototype as required.

Note the typing of return nodes is a bit more subtle than in scalac, because
dotc trees are immutable, so the tree pointed at by the context is of no help -
it does not have a return type which we could make use of.