Simplify and fix bounds propagation in constraints.
The previous scheme did not ensure that transitivity was eliminated.
Example scenario:

We have in the constraint

   P <: Q

for constrained variables P, Q and add

   Q <: T

Previous propagation added the constraint and then tested
whether the bounds of all variables were satisfiable. For
Q we test

    P <: T

but that is true because P <: Q and we already added the constraint Q <: T.
So we fail to add the constraint P <: T, and transitivity is no longer eliminated.

Instead we now test the new bounds (in this case P <: T) *before* adding
the new constraint Q <: T. This is also simpler than the previous scheme.