Removed check from addConstraint
addConstraint contained a special case where a situation like

  P { ... } <: P

was short-pathed to `return true`. But the same was not done
if the constraint was added indirectly by propagation. It's not clear
whether a special treatement of this is needed for correctness. If
it is needed, then it would be needed eberywhere. So wince we do
not want to implement it everywhere wihtout proof of necessity,
it is better to fail fast and drop the special treatment entirely.