Avoid adding constraint under typeVarsMissContext is true
SubType check is true anyway, no need to spend work. Also this allows a more
consistent treatment of bounded wildcard types in addConstraint and isSubType. In both cases
we now compare with the outer bound of the wildcard type. Previously addConstraint treated
boundes wildcard types like unbounded ones, which is not consistent.