Fixed soundness hole in TypeComparer
Two named types with same names and NoPrefix prefixes are not necessarily equal!
The fix uncovered an error in tailrec. When run on Decorators.scala, tailrec in its old
position at the end of a group produces not -Ycheckable code. Problem was fixed
by moving TailRec into its own group.