Plugging a possible hole in unify
unify now does a full `replace` of one parameter with the other. The previous
scheme of just updating the entry of a parameter risks breaking constraint
invariants. In particular, other constraint bounds can still refer to the eliminated
parameter. This might not be a problem, but then we should be systeamtically leaving
these references in everywhere. In any case it seems better to harmonize what unify
and replace are doing.