SI-8930 - Vector updated, +:, and :+ slow when typed as Seq[A]
Vector was intercepting only the IndexedSeq CanBuildFrom to quickly generate new vectors.  Now it intercepts immutable.Seq and collection.Seq as well.

There are other possibilities (collection.IndexedSeq), but they will probably arise rarely, and to avoid an absurdly long set of checks we would need a marker trait (that is not binary compatible).