Revert following underlyingIfRepeated when matching
This was a left-over from a failed attempt to have OverridingPiars work exclusively by comparing signatures.
If we do that then repeated and underlying do have the same signature and therefore are supposed to match.
But as @retronym notes, this leads to problems. In any case, we no longer try to make overriding pairs
work that way, because it fails for other reasons as well.