Added test to TreeChecker that guards against orphan parameters.
Currently, tests fail. The failures I checked are all related to tailcalls.
Not sure whether there are others.

This is a blocker for serialization. Orphan parameters cannot be serialized.

Maybe rethink the position of tailcalls? It looks to me that the repeated trouble it gives us is
more than the effort required to put an efficient tailcall recognition after pattern matching in place.
But I might be wrong.