Change scheme of translating array creations new Array(...)
The previous scheme generated too many newGenericArray expressions because
at the time newArray was called, the type arguments were not yet determined.
Furthermore, the type variables somehow did not have the right positions,
which caused them not to be interpolated and led to orphan PolyParams.

The new scheme converts the expression when the length parameter has been supplied
and it fully determines the array type before converting.