Make some tree fields lazy
Lazy fields are
 - the rhs field of a ValDef or DefDef
 - the body field of a Template

These can be instantiated with Lazy instances.
The scheme is such that lazy fields are completely
transparent for users of the Trees API.

The only downside is that the parameter used to initialize
a potentially lazy field has a weak type (now it's Any, with
Dotty it would be a union type of the form `T | Lazy[T]`.
Therefore, the parameter cannot be recovered through pattern
matching.