Changes to pickling annotations
Problem is that we need to be strict in the annotation symbol
(need to force at the latest when checking whether a definition
has some specific class of annotation), but should be lazy
in the rest of the tree.

This is achieved by pickling annotations with a symbol and a full
tree.

At the same time, Annotated trees can be ignored because they
are eliminated by Typer. We only need AnnotatedTypes.