Fix for #371
This causes backend to fail with ambiguous reference when loading
member "equals" of object "BoxesRuntime".

The reference is indeed ambiguous. BoxesRunTime has a two paremeter
equals method and, with the fix, inherits the one-parameter method from
Object. The backend needs to disambiguate, e.g. by demanding the `decl`
equals in BoxesRunTime instead of the member.