Fix variance of type prefixes in TypeMap and TypeAccumulator
According to the "new" interpretation of T#U as \exists x: T. x.U,
x.U is not a subtype of T#U. Hence variance in prefixes of named
types should be 0.

There is one case where this leads to problems. (Typing failure in
Desugar.scala, to be exact). This is when computing variances
in local type inference. We solve this problem locally
by overriding accumulator application to prefixes in the specific
variances accumulator. Because the "variances" accumulator is now
irregular wrt to the rest of variance handling, we move it
from Types to Inferencing, which is the only place where it is needed.