Avoiding dependent method types in closures
The previous logic for avoiding dependent method types in closures had a hole.
The problem arose when the expected return type of a closure was a type variable.
Then, the that type variable would be taken as the declared result type of the
closure without (at first) checking the body. The type variable would not yet
be bounded and therefore would not represent a dependent method type. Afterwards
when typechecking the closure the type variable woul dbe bounded and instantiated.
But at that time, all checking and possibly avoiding of depenencies has already
happened.

We solve the problem by typing the body of a closure during Namer whenever the expected
type is not fully defined.

Doing so uncovered a problem that anonymus function were seen as the target of returns
(tehy shoul be skipped instead). Thsi problem is fixed by a patch to SymDenotations#isSourceMethod.