Keeping track of unpickling definition order in Pickler.
The Unpickler visits nodes in a certain order and (so far) expects to see
definitions before uses. This commit makes sure that the Unpickler generates
definitions in an order that matches the pickler's behavior. Every definition
should be "pre-registered" before it can be used. This still allows for mutual
recursion between symbols because preregistering enters all symbols of a scope
in bulk before generating any references to these symbols.

It would be nice if this was the end of the story, but unfortunately that's not the case.
It turned out that dotc produced references that were not legal according to the
implemented model, and that are also not legal in the formal type system. Each of these
violations referred to a symbol outside its scope. There were two sources:

1) Pattern-bound variables were retained in the type of a case block and, consequently,
the type of the enclosing match. This has been fixed by a previous commit a0f8ec21c9ce5381bea780e7be89286653fc676e.

2) Dependent anonymous functions led to (illegal) dependent closures with references
to their parameters leaking out in the environment. This has been mostly fixed by the previous commits,
in particular 1c70842036b083652c3eeab83aad0b2490674bfe. But there are still two problems remaining, see:
89c8bd8a1eb9fb3f0f09f25bedb68de1ef2e2ae8.

We might fix the two problems. But it's inconceivable to me that scalac could also
produce only "hygienic" trees that do not have escaping references. There are too many situations
where we know this is not true (existential skolems, for a start).

So we choose to flag escaping variables in logs instead of treating them as errors, and to deal with the
situation in the Unpickler.