Fix tricky problem with re-establishing denotations of selftypes
Selftypes can be nastily recursive as this example of extmethods shows:

     class T[A, This <: That1[A]] extends AnyVal { self: This => ...

The problem, of course, is that This is really `this.This` and to determine
what that is we look up the self type which leads us right back to where we were.
We fix the problem by special casing references to class members in self-types.
These are always written symbolically, and compelmentary logic in unpickling
makes sure we do not need to compute the denotation of such members, but
instead take the denotation directly from the symbol.