Deal gracefully with out-of-scope references wehn unpickling
Out-of scope references are evil but do arise. Dotc currently produces them
in two cases (t1957.scala and Typer.scala) because it does not correctly
hygenize dependent methods used as closurs. I beleive scalac wil generate them
in more instances.

The changes can produce dependent method types used in closures. Essentially the
problem was that the depndency was not recognized in Typers when the closure
was first created, because it badly interfered with type inference. But when unpickling
the same closure the dependency is recognized. It's too late to fail now, we better
deal with this gracefully. That's why there is now a mode bit "AllowDependentFunctions"
which should be turned on only for unpickling, which reconstitutes dependent functions
as closures without complaining.