add/strict
Add -strict option to  do some type checks that are encessary to ensure type soundness, but
are stricter than what Scala 2.x enforces.

The first such test is the "pattern cannot be uniquely instantiated" problem where we reject a
non-variant case subclass of a covariant superclass in a pattern match. The error is now only
issued in -struct mode, otherwise it will be a warning.

We might move more tests into the same category. This should help the transition.