SI-8254  List SerializationProxy fails to default(Read/Write)Object
Added `defaultWriteObject` to the beginning of `writeObject` and `defaultReadObject` to the beginning of `readObject` as required by specs:
[writing](http://docs.oracle.com/javase/6/docs/platform/serialization/spec/output.html#861), [reading](http://docs.oracle.com/javase/6/docs/platform/serialization/spec/input.html#2971).

Verified that it is a no-op in terms of serialization stream (but it provides hooks that Infinispan and others may use).

No explicit tests.  If there is a change in serialization, t8549 will catch it.