Use invalidateInheritedSymbols instead of syncWithParents
syncWithParents explores the denotations of symbols that might yet to be entered
in the current run. If such a symbol is looked at in a new run before a new one
is entered, the validty period of the old denotation is extended to the new run
and consequently references to that symbol do not know they need to reload. This
pollutes the cache of references and causes StaleSymbol errors down the line.

Replacing with invalidateInheritedSymbols avoids the problem.