Test stillValid at initial phase, revive -twice
stillValid should always be tested at the phase where the initial
denotation is defined. If we do not do this we get false stale symbol errors.
Here is a scenario:

To `bringForward` `Predef$$ArrowAsscoc`, we check whether its initial denotation
has an owner which contains the denotation as one of its members. But if we do this
at a later phase (e.g. GenBCode), the owner of the initial denotation is Predef, but
Predef no longer contains ArraoAssoc - it lost the member at flatten.

With this fix we can run everything with -twice, except dotc_core. I left a comment what
goes wrong there.