Allow change of symbols in WithFixedSym#newLikeThis
The symbol might be overridden, in which case the overriding
symbol should be chosen.

This was observered when trying to pickle test all of ast.
Compiling Trees and untpd together gave discrepancies involving
`untpd.dotty$tools$dotc$ast$Trees$$Instance$$T`. It was
found that the symbol referred to by that type refers
to the abstract type in Trees.Instance, not the alias in
untpd. The bad type was produced by an asSeenFrom of the
`WithFixedSym` type
`Instance.this.untpd.dotty$tools$dotc$ast$Trees$$Instance$$T`
as seen from `untpd`. The as seen from did not pick the new
instance in `untpd`.