Fixes to double binding check.
There was a spurious error when doing pickling tests with -YnoDoubleBindings
The test manifested itself when compiling `Trees` and `untpd`. The problem was
a sourious double assignment to the `untpd$` TypeRef. The problem was that
the old `untpd` symbol did get a runId of 3, because it was looked at immediately
before being overritten with the new symbol. So the system detected a race condition.

The new test looks at the denotations previous run id instead of the symbol's when
setting a denotation. That runId was still 2, so not real double binding occurred.